+++ src/main/java/com/ll/gooHaeYu/domain/jobPost/jobPost/repository/QuestionItemRepository.java
@@ -1,6 +1,6 @@
-package com.ll.gooHaeYu.domain.jobPost.questionItem.repository;
+package com.ll.gooHaeYu.domain.jobPost.jobPost.repository;
 
-import com.ll.gooHaeYu.domain.jobPost.questionItem.entity.QuestionItem;
+import com.ll.gooHaeYu.domain.jobPost.jobPost.entity.QuestionItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
