--- src/main/java/com/ll/gooHaeYu/domain/jobPost/questionItem/controller/QuestionItemController.java
@@ -1,9 +1,0 @@
-package com.ll.gooHaeYu.domain.jobPost.questionItem.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class QuestionItemController {
-}
