--- src/main/java/com/ll/gooHaeYu/domain/jobPost/questionItem/service/QuestionItemService.java
@@ -1,11 +1,0 @@
-package com.ll.gooHaeYu.domain.jobPost.questionItem.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class QuestionItemService {
-}
